# `requestResponseMessage` handlers optimization