@Tsury 2023-12-22
> @Tsury 👏 What is different between this PR and my PR? The only different I notice is that you use a global map instance whereas I scoped the map instance to the wrapped created Proxy.

👋 I'm actually not entirely sure. Maybe the scoped usage just moves the optimization issue from the function creation to the map creation? 



> One other approach I explored was only making one map per endpoint instance: [foxglove@be58259](https://github.com/foxglove/comlink/commit/be58259d8998966ac2b885b31256ba333ea58ed5)

So this exact commit is a working, different approach? I can test it once I get back home to see how it performs.

> I wonder if your usage pattern is such that you are making lots of `createProxy` calls?

Not at all, just 3 `createProxy` executions exist in my test.

> It would be good to understand the nuances and also whether it seems "safe" to have one global map. One _gotcha_ that comes to mind with the global map (and possibly even the local map) is holding onto resolve functions after an endpoint is torn down (potentially via .terminate() calls). The current behavior with "addEventListener" will teardown all of the handlers when the endpoint itself is gone since that's managed by the underlying system but with manual listener management more care needs to be taken or at least we need to think through whether it is possible to fail in such a way.

I agree. We need to ensure that the managed approach does not introduce any issues due to dangling/missing listeners. I'll try to think of nuanced scenarios to test, appreciate any help with this - the more people tackling this the better.