@lvivski 2024-01-16
Thank you @defunctzombie and @Tsury for looking into this. Our project has ComLink bottlenecks in event listeners being added/removed all the time.

@surma @benjamind is it something that you might be interested in optimizing?