@defunctzombie 2024-01-16
I have not forgotten about this and we at Foxglove are looking at benchmarking this and working up an implementation that will meet the desired criteria.