@achim-k 2024-01-18
I have compared the three different versions (this PR, #649 and https://github.com/foxglove/comlink/commit/be58259d8998966ac2b885b31256ba333ea58ed5) with the current released comlink version and I can confirm that performance is overall better.

https://github.com/foxglove/comlink/commit/be58259d8 perform best, mainly due to the [simplification](https://github.com/foxglove/comlink/blob/be58259d8/src/comlink.ts#L644) of how the random request id is generated. Without that, it is basically the same as this PR. For all three versions, the speedup is especially noticeable when increasing the number of parallel requests (-> more pending listeners). If there is mostly just one request at a time, then the performance improvement is only marginal. The table below shows the median time in ms for repeatedly (10000 times) calling a void function exposed by the worker:

| parallel requests | current comlink | this PR           | https://github.com/foxglove/comlink/commit/be58259d8 | #649              |
|-------------------|-----------------|-------------------|------------------------------------------------------|-------------------|
| 1                 | 842             | 788 (6% faster)   | 770 (8% faster)                                      | 823 (2% faster)   |
| 2                 | 1074            | 929 (13% faster)  | 827 (23% faster)                                     | 1016 (5% faster)  |
| 5                 | 2085            | 1507 (28% faster) | 1393 (33% faster)                                    | 1647 (21% faster) |

Note that for #649 it makes a huge difference if the function proxy is cached like shown below:

```js
const obj = Comlink.wrap(worker);
const someFunctionProxy = obj.someFunction;
const numParallelRequest = 5;
for (let i = 0; i < 10_000; i++) {
  const promises = new Array(numParallelRequest).fill().map(() => someFunctionProxy());
  await Promise.all(promises);
}
```

If that is not done, a new proxy is created every time by calling `await obj.someFunction()`, which causes #649 to perform significantly worse than the current released version.

To move forward, I would suggest to 
- split out the random ID generation from #649 and create a seperate PR for it
- create a PR for https://github.com/foxglove/comlink/commit/be58259d8 which seems superior over this PR in terms of cleaning up resolve functions